backend/requirements.txt
flask
flask-cors
pycryptodome
pqcrypto
cryptography
